Memory is now stored in a Map instead of an Array.
The original intention with this project was to utilize immutable values as much as possible. Array was settled upon due to its random access speed, but arrays in F# are mutable. After some thought I've changed everything over to Map, an immutable collection, which honestly shouldn't be too much of a performance hit due to the max size being only 65535 entries.